CloudSave_v2
MOVE increase/decrease based on a CommandBatchSystem. But still not decrease when not enough MOVE currency

+++ Assets/[MAIN]LOT_Turnbase/UI/Scripts/MainUI.cs
@@ -6,7 +6,7 @@ namespace JumpeeIsland
     public class MainUI : Singleton<MainUI>
     {
         [NonSerialized] public UnityEvent<IGetCreatureInfo> OnShowCreatureInfo = new(); // send to UIManager; invoke at Units
-        [NonSerialized] public UnityEvent<int> OnRemainStep = new(); // send to UIManager; invoke at UnitManagers
+        [NonSerialized] public UnityEvent<long> OnRemainStep = new(); // send to UIManager; invoke at UnitManagers
         [NonSerialized] public UnityEvent OnClickIdleButton = new(); // send to PlayerFactionManager; invoke at IdleButton & MovingPath
         [NonSerialized] public UnityEvent<int> OnGameOver = new(); // send to GameOverAnnouncer; invoke at PlayerFactionManager
     }
